requirements.txt
beautifulsoup4==4.11.1
PyYAML==6.0
requests==2.27.1
rsa==4.8
